TST: Reorganizing qc test module.
